Added loading arena file and created new Arena folder
