+++ Arena/Layouts/arena1.txt
@@ -1,0 +1,13 @@
+wsswwwwwmmggggg
+wwwwwwggggggggg
+ggggggggggggggg
+gggggggggggssgg
+gggggggggggssgg
+ggggggggggggggg
+ggggggmmggggggg
+gggggmwwmgggggg
+gggggmwwmgggggg
+gggggmwwmgggggg
+gggggmwwmgggggg
+gggggmwwmgggggg
+gggggmwwmgggggg
